# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marcaê API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: barbershops
  - name: settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthStatus" }

  /auth/register-shop:
    post:
      operationId: registerShop
      tags: [auth]
      summary: Register a new barbershop and admin account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterShopInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400":
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Slug already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /auth/register-client:
    post:
      operationId: registerClient
      tags: [auth]
      summary: Register a new client account in an existing barbershop
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterClientInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400":
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Barbershop not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (slug + email + password)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout (invalidate current session)
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user + barbershop + plan status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /barbershop:
    patch:
      operationId: updateBarbershopSettings
      tags: [settings]
      summary: Update the authenticated admin's establishment settings (name, brand colors, contact)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBarbershopInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Barbershop" }
        "400":
          description: Invalid input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Admin only
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  /barbershops/{slug}/exists:
    get:
      operationId: checkBarbershopSlug
      tags: [barbershops]
      summary: Check whether a barbershop slug exists (used by client signup flow)
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BarbershopLookup" }

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]

    RegisterShopInput:
      type: object
      properties:
        slug: { type: string, minLength: 2 }
        name: { type: string, minLength: 1 }
        ownerName: { type: string, minLength: 1 }
        ownerEmail: { type: string, format: email }
        password: { type: string, minLength: 6 }
        phone: { type: string }
      required: [slug, name, ownerName, ownerEmail, password]

    RegisterClientInput:
      type: object
      properties:
        slug: { type: string, minLength: 2 }
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        phone: { type: string }
        password: { type: string, minLength: 6 }
      required: [slug, name, email, phone, password]

    LoginInput:
      type: object
      properties:
        slug: { type: string, minLength: 2 }
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [slug, email, password]

    Barbershop:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        ownerName: { type: string }
        ownerEmail: { type: string }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        plan:
          type: string
          enum: [trial, base, medio, super, premium, pending, expired]
        trialEndsAt: { type: string, format: date-time }
        subscriptionRenewsAt: { type: string, format: date-time, nullable: true }
        brandPrimary: { type: string, description: "Primary brand color (hex #RRGGBB)" }
        brandAccent: { type: string, description: "Accent / dark brand color (hex #RRGGBB)" }
        bookingBufferMinutes: { type: integer, minimum: 0 }
        bookingAvailabilityMode:
          type: string
          enum: [duration_buffer, release_on_complete]
        businessSchedule:
          $ref: "#/components/schemas/BusinessSchedule"
        intakeFields:
          type: array
          items: { $ref: "#/components/schemas/IntakeField" }
        createdAt: { type: string, format: date-time }
      required: [id, slug, name, ownerName, ownerEmail, plan, trialEndsAt, brandPrimary, brandAccent, bookingBufferMinutes, bookingAvailabilityMode, businessSchedule, intakeFields, createdAt]

    UpdateBarbershopInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        brandPrimary:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          description: "Primary brand color (hex #RRGGBB)"
        brandAccent:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          description: "Accent / dark brand color (hex #RRGGBB)"
        bookingBufferMinutes: { type: integer, minimum: 0 }
        bookingAvailabilityMode:
          type: string
          enum: [duration_buffer, release_on_complete]
        businessSchedule:
          $ref: "#/components/schemas/BusinessSchedule"
        intakeFields:
          type: array
          items: { $ref: "#/components/schemas/IntakeField" }

    BusinessSchedule:
      type: object
      additionalProperties:
        type: object
        properties:
          enabled: { type: boolean }
          startTime: { type: string }
          endTime: { type: string }
        required: [enabled, startTime, endTime]

    IntakeField:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        type:
          type: string
          enum: [text, textarea, date, phone]
      required: [key, label, type]

    BarbershopLookup:
      type: object
      properties:
        exists: { type: boolean }
        name: { type: string, nullable: true }
        brandPrimary: { type: string, nullable: true }
        brandAccent: { type: string, nullable: true }
      required: [exists]

    AuthUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        barbershopId: { type: string, format: uuid }
        role:
          type: string
          enum: [admin, employee, client]
        name: { type: string }
        email: { type: string }
        phone: { type: string, nullable: true }
        professionalId: { type: string, format: uuid, nullable: true }
        clientId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, barbershopId, role, name, email, createdAt]

    PlanStatus:
      type: object
      properties:
        plan:
          type: string
          enum: [trial, base, medio, super, premium, pending, expired]
        trialDaysLeft: { type: integer }
        isActive: { type: boolean }
        isPremium: { type: boolean }
        isPaid: { type: boolean }
        planName: { type: string }
        planPrice: { type: string, nullable: true }
        trialEndsAt: { type: string, format: date-time }
      required: [plan, trialDaysLeft, isActive, isPremium, isPaid, planName, planPrice, trialEndsAt]

    AuthSession:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/AuthUser" }
        barbershop: { $ref: "#/components/schemas/Barbershop" }
        planStatus: { $ref: "#/components/schemas/PlanStatus" }
      required: [token, user, barbershop, planStatus]